requirements.txt
streamlit
pandas
openpyxl
plotly
numpy
pillow
openai
python-dotenv